+++ tests/test_popover.py
@@ -10,6 +10,7 @@ def test_001po_popover(dash_duo):
     app.layout = dmc.MantineProvider(
         html.Div(
             [
+                html.Div("Click Outside", id="click-outside", style={"background": "grey"}),
                 dmc.Popover(
                     [
                         dmc.PopoverTarget(dmc.Button("Toggle Popover", id="btn")),
@@ -32,13 +33,30 @@ def test_001po_popover(dash_duo):
     # Wait for the app to load
     dash_duo.wait_for_text_to_equal("#output", "False")
 
+    # Open Popover
     popover_btn = dash_duo.find_element("#btn")
     popover_btn.click()
 
-    # Verify the output
+    # Verify the Popover opens when clicking the target
     dash_duo.wait_for_text_to_equal("#output", "True")
 
+    # Close Popover
     popover_btn.click()
+
+    # Verify the Popover closes when clicking the target
+    dash_duo.wait_for_text_to_equal("#output", "False")
+
+    #Open Popover
+    popover_btn.click()
+
+    # Verify the Popover is open
+    dash_duo.wait_for_text_to_equal("#output", "True")
+
+    # Click outside
+    click_outside = dash_duo.find_element("#click-outside")
+    click_outside.click()
+
+    # Verify the Popover closes when clicking outside
     dash_duo.wait_for_text_to_equal("#output", "False")
 
     assert dash_duo.get_logs() == []